@@ -10,10 +10,11 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-import os
 import sys
+from pathlib import Path
 
-sys.path.insert(0, os.path.abspath("./../.."))
+docs_dir = Path(__file__).parent.parent
+sys.path.insert(0, docs_dir.resolve().as_posix())
 
 
 # -- Project information -----------------------------------------------------
