@@ -4,7 +4,7 @@ a Telegram bot using the Pyrogram library.
 """
 
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2022 Victor Cebarros <https://github.com/victorcebarros>
+# Copyright (c) 2023 Victor Cebarros <https://github.com/victorcebarros>
 
 __author__: str = "Victor Cebarros"
 __version__: str = "0.1.0"
