@@ -8,9 +8,10 @@ All Korone's constants that are used throughout the codebase.
 """
 
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2022 Victor Cebarros <https://github.com/victorcebarros>
+# Copyright (c) 2023 Victor Cebarros <https://github.com/victorcebarros>
 
 import os
+from pathlib import Path
 
 LOGGER_FORMAT_OUTPUT: str = (
     "%(name)s"
@@ -23,22 +24,22 @@ LOGGER_FORMAT_OUTPUT: str = (
 )
 """The format to be used in the logger."""
 
-XDG_CONFIG_HOME: str = os.environ.get("XDG_CONFIG_HOME", "~/.config")
+XDG_CONFIG_HOME: str = os.environ.get("XDG_CONFIG_HOME", str(Path(Path.home() / ".config")))
 """
 The XDG_CONFIG_HOME environment variable.
 Where user-specific configurations should be written (analogous to /etc).
 """
 
-XDG_DATA_HOME: str = os.environ.get("XDG_DATA_HOME", "~/.local/share")
+XDG_DATA_HOME: str = os.environ.get("XDG_DATA_HOME", str(Path(Path.home() / ".local/share")))
 """
 The XDG_DATA_HOME environment variable.
 Where user-specific data files should be written (analogous to /usr/share).
 """
 
-DEFAULT_CONFIG_PATH: str = f"{XDG_CONFIG_HOME}/korone/korone.conf"
+DEFAULT_CONFIG_PATH: Path = Path(f"{XDG_CONFIG_HOME}/korone/korone.conf")
 """The default path to the config file."""
 
-DEFAULT_DBFILE_PATH: str = f"{XDG_DATA_HOME}/korone/korone.db"
+DEFAULT_DBFILE_PATH: Path = Path(f"{XDG_DATA_HOME}/korone/korone.db")
 """The default path to the database file."""
 
 DEFAULT_WORKERS: int = 24
@@ -49,26 +50,26 @@ DEFAULT_NAME: str = "korone"
 
 DATABASE_SETUP: str = """\
 CREATE TABLE IF NOT EXISTS Users (
-    uuid INTEGER PRIMARY KEY,
+    id INTEGER PRIMARY KEY,
     language VARCHAR(2) NOT NULL DEFAULT "en",
     registrydate INTEGER NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS Chats (
-    uuid INTEGER PRIMARY KEY,
+    id INTEGER PRIMARY KEY,
     language VARCHAR(2) NOT NULL DEFAULT "en",
     registrydate INTEGER NOT NULL,
     chattype TEXT NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS DisabledCommands (
-    chat_uuid INTEGER,
+    chat_id INTEGER,
     command TEXT,
     state BIT
 );
 
 CREATE TABLE IF NOT EXISTS Filters (
-    chat_uuid INTEGER,
+    chat_id INTEGER,
     handler TEXT,
     data TEXT,
     file_id TEXT,
