@@ -3,13 +3,13 @@ Entry point of the Korone.
 """
 
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2022 Victor Cebarros <https://github.com/victorcebarros>
+# Copyright (c) 2023 Victor Cebarros <https://github.com/victorcebarros>
 
 import logging
 
 from korone import config
+from korone.database.impl.sqlite3_impl import SQLite3Connection
 from korone.modules import App, AppParameters
-from korone.database import Database
 
 log = logging.getLogger(__name__)
 
@@ -27,24 +27,27 @@ def main(argv: list[str]) -> int:
     """
     log.info("Program started")
 
-    config.init("korone.conf")
+    config.init()
 
     ipv6 = config.get("pyrogram", "USE_IPV6").lower() in ("yes", "true", "1")
 
-    Database.connect("korone.db")
-    Database.setup()
+    with SQLite3Connection() as conn:
+        if not conn:
+            log.error("Database connection failed")
+            return 1
 
-    param: AppParameters = AppParameters(
-        api_id=config.get("pyrogram", "API_ID"),
-        api_hash=config.get("pyrogram", "API_HASH"),
-        bot_token=config.get("pyrogram", "BOT_TOKEN"),
-        ipv6=ipv6,
-    )
+        conn.setup()
 
-    app: App = App(param)
-    app.setup()
-    app.run()
+        param: AppParameters = AppParameters(
+            api_id=config.get("pyrogram", "API_ID"),
+            api_hash=config.get("pyrogram", "API_HASH"),
+            bot_token=config.get("pyrogram", "BOT_TOKEN"),
+            ipv6=ipv6,
+            connection=conn,
+        )
 
-    Database.close()
+        app: App = App(param)
+        app.setup()
+        app.run()
 
     return len(argv) - 1
