@@ -3,15 +3,15 @@ The toggle module enables and disables commands on a per chat basis.
 """
 
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2022 Victor Cebarros <https://github.com/victorcebarros>
+# Copyright (c) 2023 Victor Cebarros <https://github.com/victorcebarros>
 
 from pyrogram import Client, filters
 from pyrogram.types import Message
 
-from korone.modules.hello import get_language_code
-from korone.modules.core import toggle
-from korone.database.manager import Command
+from korone.database.sqlite3_manager import Command
 from korone.locale import StringResource
+from korone.modules.core import toggle
+from korone.modules.hello import get_language_code
 from korone.utils.misc import get_command_arg
 
 
@@ -29,9 +29,7 @@ async def command_disable(_, message: Message) -> None:
 
     if command == "":
         await message.reply(
-            StringResource.get(
-                language_code, "strings/enable/message/failure/emptycommand"
-            )
+            StringResource.get(language_code, "strings/enable/message/failure/emptycommand")
         )
         return
 
@@ -45,9 +43,7 @@ async def command_disable(_, message: Message) -> None:
         )
         return
 
-    await message.reply(
-        StringResource.get(language_code, "strings/disable/message/success")
-    )
+    await message.reply(StringResource.get(language_code, "strings/disable/message/success"))
 
 
 @Client.on_message(filters.command("enable"))
@@ -64,9 +60,7 @@ async def command_enable(_, message: Message) -> None:
 
     if command == "":
         await message.reply(
-            StringResource.get(
-                language_code, "strings/enable/message/failure/emptycommand"
-            )
+            StringResource.get(language_code, "strings/enable/message/failure/emptycommand")
         )
         return
 
@@ -80,6 +74,4 @@ async def command_enable(_, message: Message) -> None:
         )
         return
 
-    await message.reply(
-        StringResource.get(language_code, "strings/enable/message/success")
-    )
+    await message.reply(StringResource.get(language_code, "strings/enable/message/success"))
